Send rendered actions when they're made on the rendered side.
This required differentiating between the different surface types and between a transform for the surface vs for the observation.

This fixes the case of rendering rgb without feature layers, and a deadlock when re-initializing the window.

Fix the help screen to have its size calculated based on the contents so it fits in the smaller window of rgb only.

PiperOrigin-RevId: 177040399